pyproject.toml
[tool.poetry]
name = "homework drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
setuptools = "^72.0.0"
stripe = "^10.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
